feat: move GenericLoadable and UrlPreviewState sealed classes to commons
Move GenericLoadable and UrlPreviewState from amethyst.ui.components
to amethyst.commons.ui for KMP iOS availability.

Kotlin typealiases for sealed classes with nested types break `is`
checks and smart casts (e.g. `is GenericLoadable.Loaded`). By moving
these sealed classes to commons and updating all 26 consumer files to
import directly from commons, we avoid this issue entirely. The old
locations retain typealiases for backward compatibility.

Part of KMP iOS migration (#2238).
