+++ amethyst/src/main/java/com/vitorpamplona/amethyst/service/cashu/CachedCashuParser.kt
@@ -21,7 +21,7 @@
 package com.vitorpamplona.amethyst.service.cashu
 
 import android.util.LruCache
-import com.vitorpamplona.amethyst.ui.components.GenericLoadable
+import com.vitorpamplona.amethyst.commons.ui.GenericLoadable
 import kotlinx.collections.immutable.ImmutableList
 
 object CachedCashuParser {
